refactor: Update Nav component to include CSS Gradient Generator link
The Nav component has been updated to include a link to the CSS Gradient Generator page in the dropdown menu. This will provide users with a convenient way to access the gradient generator feature. The code has been refactored to add the link and adjust the styling of the dropdown menu.

Ref: #60