feat: Add JSON Generator link to Nav component
The Nav component has been updated to include a link to the JSON Generator page in the dropdown menu. This will provide users with a convenient way to access the JSON Generator feature. The code has been modified to add the link and adjust the styling of the dropdown menu.

Ref: #60