feat: Add download functionality to CardForm component
The code changes add the ability to download the generated data as a JSON file in the CardForm component of the JSON Generator feature. When the "Download" button is clicked, the data is converted to a JSON string and saved as a file using the file-saver library. This feature enhances the usability of the JSON Generator form.

Ref: #60